+++ src/orb/____inner_core/util/function/_GetZ_T.java
@@ -4,5 +4,5 @@ package orb.____inner_core.util.function;
  * Created by doekewartena on 9/11/15.
  */
 public interface _GetZ_T<T> {
-    double getZ(T t);
+    float getZ(T t);
 }
